<!-- Thank you for submitting a pull request! Find more information in our development guide at https://github.com/trinodb/trino/blob/master/.github/DEVELOPMENT.md and contact us on #dev in Slack. -->
<!-- Provide an overview of the PR for maintainers and reviewers. -->
## Description

Change defaults of `max-unacknowledged-splits-per-task` and `max-adjusted-pending-splits-per-task` to 2000 for better overall performance.

`SELECT count(*) from hive.test.lineitem_parquet_64_group`

old: 6,9304 s
new: 5,8872 s

`SELECT count(orderkey) from hive.test.lineitem_parquet_64_group`

old: 9,0065 s
new: 8,0764 s

`SELECT count(orderkey),count(suppkey)  from hive.test.lineitem_parquet_64_group`

old: 22,1207 s
new: 21,2324 s

![image](https://user-images.githubusercontent.com/6515994/210357936-8ddb31a0-5164-4d0a-a9e6-d7a3e647e460.png)

Concurrency benchmark (queries per hours):
new: 7698 
old: 7269 

<!-- Provide details that would help an engineer who is unfamiliar with this part of the code. -->
## Additional context and related issues



<!-- Mark the appropriate option with an (x). Propose a release note if you can. -->
## Release notes

(x) This is not user-visible or docs only and no release notes are required.
( ) Release notes are required, please propose a release note for me.
( ) Release notes are required, with the following suggested text: