@pettyjamesm 2023-01-04
> @pettyjamesm this PR increases unacknowledged split size to 2000, I think it's fine, wdyt?

I'm a little skeptical that raising the value to 2,000 is necessary or actually safe to pick up as the default based on the results of running `select count` queries over parquet files. It's not especially surprising that this change would generate those improvements in that specific scenario since each split will be exceedingly cheap to process and the scheduler latency will dominate query performance- but the TPCH and TPCDS queries appear to show no improvement beyond what looks to be the margin of run-to-run variance.

These configuration properties are probably rarely configured away from their defaults in most deployments so we want to be a little cautious about picking overly-aggressive defaults. In particular, increasing `max-unacknowledged-splits-per-task` by 4x could result in (almost) 4x larger task update payloads which might create problems for some users.

I have a couple recommendations about how you might want to proceed:

1. Consider re-running TPCH and TPCDS benchmarks with "small file" (maybe ~1-10MB per file?) data sets to see whether increasing the value to 1,000 or 2,000 actually shows an improvement there without over-biasing towards very "cheap" workloads like `select count(*)`

2. Consider making `max-unacknowledged-splits-per-task` a session property so that higher values can be provided in specific situations where the improvement is expected to be "safe" and the performance improvement is appreciable (eg: `select count(*)` over tables with small files and without too many columns).