@sopel39 2023-01-04
@pettyjamesm 

> These configuration properties are probably rarely configured away from their defaults in most deployments so we want to be a little cautious about picking overly-aggressive defaults. In particular, increasing max-unacknowledged-splits-per-task by 4x could result in (almost) 4x larger task update payloads which might create problems for some users.

What was the error that you've been actually observing? I'm particularly interested why 500 was chosen by default for `max-unacknowledged-splits-per-task` and not some other value.

> Consider re-running TPCH and TPCDS benchmarks with "small file" (maybe ~1-10MB per file?) data sets to see whether increasing the value to 1,000 or 2,000 actually shows an improvement there without over-biasing towards very "cheap" workloads like select count(*)

It's not really about TPCH/TPCDS but rather about workloads with:
* selective queries (data skipping)
* empty splits (big row groups)
* small splits
* cache