@pettyjamesm 2023-01-04
> What was the error that you've been actually observing?

Depends on how large the payloads get. Potentially you could trigger request timeouts because of the amount of time the worker spends parsing the task update payload JSON, but before that point you'll see high allocation and GC activity on the coordinator. In presto they added a hard limit of 16MB on the task update body and fail the task immediately to avoid some of the issues they saw at the time.

> I'm particularly interested why 500 was chosen by default for max-unacknowledged-splits-per-task and not some other value.

The value was chosen before the existence of the `QueueSizeAdjuster` in `UniformNodeSelector`. At the time it was clear that there needed to be _some_ limit on the maximum number of splits sent in a single request but not clear what that limit should be, so I experimented with the small file datasets and chose a value large enough that the performance of queries like `select sum(column) from ...` no longer improved without seeming "unreasonably high" to me personally. I chose `sum` instead of `count` to avoid over-biasing towards _super_ cheap queries and to ensure that the data would actually be read from the input files instead of just getting picked out of the parquet footer.