@Dith3r 2023-01-05
Test:

`SELECT sum(orderkey / 10000) from hive.test.lineitem_parquet_64_group`

old: 10.0148 ±0.6281
new: 9.6028 ±0.5703