@JunhyungSong 2023-01-14
The concerns that @pettyjamesm mentioned can be mitigated by trinodb#15721.