@Dith3r 2023-01-31
@pettyjamesm Adaptive request size was merged, do you think that increasing configuration values are OK now?