@Dith3r 2023-01-31
There are a few configuration options tested like max-splits-per-node, max-adjusted-pending-splits-per-task, max-unacknowledged-splits-per-task tested with range from 1000 up to 4000 with step of 500 (and others like min-pending-splits-per-task with different range) with set of simple queries. Configurations which presented best outcome were tested with tpch, tpds and concurrency test suites. Values presented in this PR presented as better overall setup. 

> there's some risk that this could increase processing time skew between workers when some splits are significantly more expensive than others.

Still there is `min-pending-splits-per-task` which is increased only if worker process splits faster than receive form coordinator. Coordinator needs to have more splits to assign, and node was marked as full. Otherwise, it works as before.